fixed a lot of stuff and changed control to backend

+++ src/app/views.py
@@ -1,11 +1,14 @@
 # A very simple Flask Hello World app for you to get started with...
 
-import subprocess, signal, os, time, json, zipfile, shutil
+import subprocess, signal, os, time, json, zipfile, shutil, datetime
 from moviepy.video.io.VideoFileClip import VideoFileClip
 
 # JJ - app is run from home directory, so we need the path to be relative to that
-#DATA_PATH = "app/static/data"
-DATA_PATH = "catkin_ws/src/web_remote/src/app/static/data"
+
+ROOT_PATH = os.path.dirname(os.path.abspath(__file__))
+STATIC_PATH = os.path.join(ROOT_PATH, "static")
+DATA_PATH = os.path.join(ROOT_PATH, "static/data")
+TEMP_PATH = os.path.join(ROOT_PATH, "static/temp")
 
 from flask import render_template, send_file, request, jsonify, current_app, request
 from app import app
@@ -64,9 +67,7 @@ def loadData():
     data = {
         'battery_percentage': 88,
         'robot_name': "turty3_pro",
-        'past_runs': [
-
-        ]
+        'past_runs': []
     }
     for folder in os.listdir(DATA_PATH):
 
@@ -84,30 +85,67 @@ def loadData():
 @app.route('/startNavigation/')
 def start_nav():
     print("START NAV")
+    current_app.config['state'] = "run"
+    current_app.config['timer'].reset()
 
-    # JJ
-    if current_app.config['controller'].get_paused() == True:
-        print("Robot controller was PAUSED, unpausing now")
-        current_app.config['controller'].set_paused(False)
-    else:
-        print ("Robot controller was NOT PAUSED, starting explore_process now")
-        devnull = open('/dev/null', 'w')
-        #current_app.config['explore_process'] = subprocess.Popen(["./mini_turty_explore.sh"], stdout=devnull, shell=False)
-        current_app.config['explore_process'] = subprocess.Popen(["./mini_turty_explore.sh", "record"], shell=False)
-
+    if not current_app.config['NO_ROBOT']:
+        # JJ
+        if current_app.config['controller'].get_paused() == True:
+            print("Robot controller was PAUSED, unpausing now")
+            current_app.config['controller'].set_paused(False)
+        else:
+            print ("Robot controller was NOT PAUSED, starting explore_process now")
+            devnull = open('/dev/null', 'w')
+            #current_app.config['explore_process'] = subprocess.Popen(["./mini_turty_explore.sh"], stdout=devnull, shell=False)
+            current_app.config['explore_process'] = subprocess.Popen(["./mini_turty_explore.sh", "record"], shell=False)
     return jsonify({})
     
 @app.route('/pauseNavigation/')
 def pause_nav():
     print("PAUSE NAV")
-
-    # JJ
-    current_app.config['controller'].set_paused(True)
+    current_app.config['timer'].pause()
+    current_app.config['state'] = "pause"
+    if not current_app.config['NO_ROBOT']:
+        # JJ
+        current_app.config['controller'].set_paused(True)
 
     return jsonify({})
 
+@app.route('/resumeNavigation/')
+def resume_nav():
+    current_app.config['timer'].resume()
+    current_app.config['state'] = "run"
+    print("RESUME NAV!")
+    return jsonify({})
+
+@app.route('/navigationStatus/')
+def nav_Status():
+    loading = True
+    run_ended = False
+    loading_msg = "Warming up the Robot..."
+
+    if not current_app.config['NO_ROBOT']:
+        #robot control code goes here
+        pass
+
+    if run_ended:
+        current_app.config['state'] = "end"
+
+
+
+    return jsonify({
+        "state": current_app.config['state'],
+        "loading": loading,
+        "run_ended": run_ended,
+        "loading_msg": loading_msg,
+        "timerStarted": current_app.config['timer'].startTimeStamp(),
+        "deltaTime": current_app.config['timer'].get_current_time(),
+    })
+
 @app.route('/saveNavigation/')
 def save_nav():
+    current_app.config['timer'].stop()
+    current_app.config['state'] = "start"
     location = request.args.get("location")
     print("SAVE NAV", location)
     
@@ -119,16 +157,16 @@ def save_nav():
     #wait for the file to appear in the /static/temp
 # JJ - app is run from home directory, so we need the path to be relative to that
 #    TEMP_PATH = "app/static/temp/latestRun"
-    TEMP_PATH = "catkin_ws/src/web_remote/src/app/static/temp/latestRun"
+    LATEST_RUN_PATH = os.path.join(TEMP_PATH, "latestRun")
     searching = True
     timeout = 0
     while searching:
-        if "video.mp4" in os.listdir(TEMP_PATH):
+        if "video.mp4" in os.listdir(LATEST_RUN_PATH):
             '''
             JJ
-            file = VideoFileClip(TEMP_PATH + "/video.mp4")
+            file = VideoFileClip(LATEST_RUN_PATH + "/video.mp4")
             data = {"runningTime": file.duration}
-            with open(TEMP_PATH + '/info.json', 'w') as f:
+            with open(LATEST_RUN_PATH + '/info.json', 'w') as f:
                 json.dump(data, f)
             
             #close moviePy file
@@ -137,25 +175,25 @@ def save_nav():
             
             #copy data to new directory
             print("Copying data...")
-            shutil.copytree(TEMP_PATH, DATA_PATH + "/" + location)
+            shutil.copytree(LATEST_RUN_PATH, DATA_PATH + "/" + location)
             #delete files from temp folder
-            for file in os.listdir(TEMP_PATH):
-                os.unlink(TEMP_PATH + "/" + file)
+            for file in os.listdir(LATEST_RUN_PATH):
+                os.unlink(LATEST_RUN_PATH + "/" + file)
             
             '''
             data = {"runningTime": 15.00}
-            with open(TEMP_PATH + '/info.json', 'w') as f:
+            with open(LATEST_RUN_PATH + '/info.json', 'w') as f:
                 json.dump(data, f)
 
             #copy data to new directory
             print("Copying data...")
-            shutil.copytree(TEMP_PATH, DATA_PATH + "/" + location)
+            shutil.copytree(LATEST_RUN_PATH, DATA_PATH + "/" + location)
             #delete files from temp folder
-            for file in os.listdir(TEMP_PATH):
-                os.unlink(TEMP_PATH + "/" + file)
+            for file in os.listdir(LATEST_RUN_PATH):
+                os.unlink(LATEST_RUN_PATH + "/" + file)
 
             # JJ - copy the map
-            shutil.copy("catkin_ws/src/web_remote/src/app/static/img/map.jpg", DATA_PATH + "/" + location)
+            shutil.copy(STATIC_PATH +  "/img/map.jpg", DATA_PATH + "/" + location)
 
             searching = False
         else:
@@ -170,11 +208,14 @@ def save_nav():
     
 @app.route('/discardNavigation/')
 def discard_nav():
+    current_app.config['timer'].stop()
     print("DISCARD NAV")
+    current_app.config['state'] = "start"
     
     # JJ 
-    current_app.config['controller'].set_paused(False)
-    kill_explore()
+    if not current_app.config['NO_ROBOT']:
+        current_app.config['controller'].set_paused(False)
+        kill_explore()
 
     return jsonify({})
 
@@ -211,10 +252,11 @@ def latestMap():
 
 # JJ
 def kill_explore ():
-    pid = current_app.config['explore_process'].pid
-    os.kill(pid, signal.SIGINT)
-    if not current_app.config['explore_process'].poll():
-        print ("Process correctly halted")
+    if not current_app.config['NO_ROBOT']:
+        pid = current_app.config['explore_process'].pid
+        os.kill(pid, signal.SIGINT)
+        if not current_app.config['explore_process'].poll():
+            print ("Process correctly halted")
 
 if __name__ == '__main__':
     raise Exception('test')
